feat: Replace 'author' with 'origin' in Plugin model
