+++ myauth/models.py
@@ -164,20 +164,21 @@ class Plugin(models.Model):
 
     id: int
     team = models.ForeignKey(Team, on_delete=models.CASCADE)  # cannot be edited by user
-    type = models.CharField(max_length=20, choices=TYPE) # cannot be edited by user
+    type = models.CharField(max_length=20, choices=TYPE)  # cannot be edited by user
     url = models.URLField(blank=False, null=False, max_length=200)  # cannot be edited by user
-    author = models.CharField(blank=False, null=False, max_length=50)  # cannot be edited by user
+    origin = models.ForeignKey("self", blank=True, null=True, on_delete=models.CASCADE)  # cannot be edited by user
     name = models.CharField(blank=False, null=False, max_length=50)
     description = models.CharField(blank=True, null=False, max_length=500)
     products = models.CharField(max_length=20, choices=PRODUCTS)
     enabled = models.BooleanField(null=False, default=False)
     collections = models.ManyToManyField(Collection, blank=True)
 
-class TrustedService(models.Model):
 
+class TrustedService(models.Model):
     id: int
     user = models.ForeignKey(User, on_delete=models.CASCADE)
     plugin = models.OneToOneField(Plugin, on_delete=models.CASCADE)
+
 
 class Usage(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
